Rearrange the navigation suggestions so they show up under their respective search queries (#52)
Change how the Suggestions List is arranged so that the navigation suggestions show up underneath the search query they're associated with.

Stop pulling out the first navigation suggestion as the "top suggestion" for now. This prevents the suggestion from showing up underneath its respective query, and we're going to be showing something else at the top of the list soon.

Try to align the styling on the navigation suggestions and the query suggestions. The mocks show that the padding on the query suggestions is only 8 vs 10 dp, but we can pixel push later.

Avoid a crash where two items in the LazyColumn are given the same key. This can happen if the exact same URL is suggested in two different sections. We'll need to revisit this when we add more things to the SuggestionsList.

Delete chips. These are obsolete and were deleted on iOS recently.

BUG=https://github.com/neevaco/neeva-android/issues/47