Address the bottom bar failing to scroll away (#80)
There appears to be a bug in the version of WebLayer that we're using, where the view fails to render when it starts to scroll away.  There is also a bad interaction with the WebLayer and Composables, where the `TabToolbar` just fails to be re-rendered even though the Composable is firing when the state changes.  To get around this, use the `BrowserControlsOffsetCallback` provided by WebLayer, which allows us to pass in a fake View that WebLayer scrolls up and down and receive offsets for how far the real bottom bar should be scrolled up or down.

Fixes: https://github.com/neevaco/neeva-android/issues/79
Fixes: https://github.com/neevaco/neeva-android/issues/44

https://user-images.githubusercontent.com/96092902/146854662-f0a17bda-8dd6-496f-bf99-2889c3ba1cd0.mp4