Fix menu handling (#94)
We should discourage having lambdas held by global objects, like the NeevaMenuData. Instead, we should be passing IDs back from the Composables and having the handlers store all the logic.

Change the AppNavModel to have explicit functions that pull up specific screens instead of passing in constants.

Change the NeevaMenuData to keep IDs instead of lambdas set by the NeevaActivity. The IDs are passed to AppNavModel.onMenuItem, which handles the logic for opening specific screens. This avoids the NeevaMenuData having to pipe logic all the way down into the Composable hierarchy used for displaying menus.

Pull out more of the Callbacks from WebLayerModel.kt and put them in their own files.

Clean up some of the logic used to restore tabs when the Activity is destroyed. This is just a cleanup -- I don't think the actual logic is correct, but I need time to investigate.