Consolidate ViewModels, and unit tests for `ActiveTabModel` and `URLBarModel` (#137)
# ViewModel changes
DomainViewModel is now subsumed by HistoryViewModel, allowing us to update both more easily.

SelectedTabModel has been renamed ActiveTabModel to go with WebLayer's terminology. It also now keeps track of and exposes the active tab Flow.

WebLayerModel now owns instances of URLBarModel, ActiveTabModel, and SuggestionsModel because they're all related, and they no longer inherit from ViewModel.

Deletes the CardViewModel, which really only existed to hoist a LazyListState out of the @Composable CardGrid. Instead, the CardsContainer now monitors for changes in the active tab and creates new LazyListState as necessary.

NeevaActivity now sets up the link between the SpaceStore refreshing from the URLBarModel to better encapsulate the classes.

WebLayerModel now sets up the link between the HistoryViewModel and the URLBarModel rather than URLBarModel knowing about it.

Add unit tests for ActiveTabModel and URLBarModel.

# Fixes
Fixes the CardGrid touch target so that you can click on the title of the card to switch to the tab.

ActiveTabModel now updates the title and URL even when the activeTab passed into onActiveTabChanged is null. This avoids situations where the bar continues to show old information.

Clicking on the URL bar when the autocomplete suggestion is either empty or has already been applied no longer hijacks the click, allowing the user to move the cursor.

# Library updates
Update to Apollo3 to get the updated Kotlin coroutine changes. This means a bunch of places that used Input and await() now use Optional and execute().

Mockito has been configured to allow mocking of final classes.