Allow more things to be injected (#144)
Get rid of NeevaBrowser.context by passing around and injecting Application instead.

HistoryViewModel is turned into an injectable HistoryManager. All functions that rely on the viewModelScope now require the callers to provide a coroutineScope. We should re-visit this once we change audit how the database works.

Consolidates tab screenshot logic in WebLayerModel, removing it from the extensions file.

Removes places that use NeevaBrowser.context as a shortcut to create defaults (thumbnail file URIs for tabs, default drawables for Spaces)

Makes WebLayerModel @Injectable so that we can get access to the Application.

Updates ApolloClient creation so that it no longer creates and stores an instance.