Put WebLayer Views into the Composable hierarchy (#568)
Add an AndroidView into the hierarchy as part of the BrowserScaffold.  This
Composable is used to house the View of the WebLayer Fragment that displays
the website.  Every time the AndroidView is recomposed, we check the current
BrowserWrapper for its View and move it into the Composable hierarchy if it
is not currently already there.

Note that because the Composables are recomposed quite often, and because the
BrowserScaffold is removed from the hierarchy whenever the user navigates to
another AppNavDestination (like the CardGrid), a new AndroidView will be created
each time and the WebLayer View will have to be moved.