Fix `IncognitoFaviconCache` stream usage (#591)
`IncognitoFaviconCache` was changed to use BitmapIO when it was introduced, but
it was saving files incorrectly: the callsite from the `FaviconCache` was using
the default FileOutputStream instead of the encrypted stream.  Because the load
was using the encrypted stream, it was failing to load the favicons later.  To
avoid this, break apart `BitmapIO` so that it's easier to parse.

* `BitmapIO` now only maintains logic for saving and loading bitmaps, without
  any support for scaling or MD5 hashing.
* `TabScreenshotManager` now calls `BitmapIO` to save and load its images.
* `SpaceStore` now owns the logic for writing scaled down spaces images.
* `FaviconCache` now owns logic for writing out MD5-hashed thumbnails.