Be consistent about dark theme check (#622)
Instead of passing around booleans that override what we think the dark theme
state is, just create a new CompositionLocal that carries the state to all the
controls.  This replaces the inconsistent boolean being passed around, as well
as the `isSystemInDarkTheme()` check.  This also opens the door to adding an
in-app setting that allows the user to override the system setting.