Allow the user to import a history database (#635)
Add the ability to import a previously exported history database for debugging.  This operation blows away the user's existing database and replace it with one that was previously exported.

This works by asking the user to select a file to import using the `ActivityResultContract` API.  Once selected, we extract the Zip file to a particular location in the `cache` directory and then kill the app.  On startup, the `HistoryDatabase` will first check to see if there's a database waiting to be imported before trying to open the database.  If a pending import is discovered, the user's `databases` directory is deleted and the imported files are copied over.

* This is a dirty way of doing it, but it works and should only happen for developers.

* We kill the process because simply asking Room to close the database didn't work.  It would see me trying to copy the files over into the `databases` directory and crash due to database corruption before the extraction was complete.

* Updates the `SettingsButton` so that it can show two lines of text.