Eliminate most compiler warnings (#650)
I'd like to get to a point where we're treating warnings as errors, which would allow us to make sure that we're not missing something actually broken.  To get there, fix a swath of warnings that are popping out at every compile.  This PR doesn't fix all of these, but follows most of the suggestions that are made.