Add public share toggle UI to native Spaces (#686)
* Add public share toggle UI to native Spaces

- Routes the share button to open a new custom share sheet UI
- Adds a public ACL toggle with the relevant info
- Adds a row of social share icons for when the toggle is enabled

This UI is not 100% done yet. We will add a list of people who have edit access
and also another section (not expanded by default) that let's you add indivisual email addresses.
But this portion is self contained enough for public Spaces.

* Tweak max lines for space name

* Moar spaces

* Fix community spaces thumbnails

Community spaces thumbnails were set to null as a byproduct of the move to url. Fixes that

* Address comments