Remove File-related StrictMode violations (#709)
Introduce a `Directories` class that asynchronously provides directories that our classes might be interested in as `Deferred<File>`s.  This avoids trying to open these files on the main thread when the app is starting up.

Callsites must now call `await()` on the `@Inject`ed `Directories` in order to get the `File` they need, which is mostly fine because any place doing I/O should have already been doing this on the IO Dispatcher.

We also convert any functions that expect to be on a `@WorkerThread` into a `suspend fun` to make it explicitly clear that something should be running on the IO thread.

After this PR, the only remaining `StrictMode` violations are `SharedPreferences` and Dagger initialization related.

Also fixes a security error with the ZipUtils usage.