Disable isolatedSplits in a debug build (#738)
This is necessary to make API level 28 happy. Apparently, the test system relies on adding a split that needs to interact in some manner with the base module. For some reason with API level 32, the test system is unperturbed by enabling isolatedSplits. It is okay however to simply disable this in debug builds since we don't package up weblayer_support as a split in debug builds.

This change also adds a script for running instrumented tests, which makes it a bit clearer what is going on by adding a log line for installing WebLayerSupport.apk and passes -r to adb install to ensure that we always use the latest version of the APK.