Fix tracking prevention toggle (#756)
This was not working properly because we were not calling stopTracking at all.
Currently we set the enableTrackingProtection to a MutableState, but MutableState doesn't update itself
value by itself outside the context of a Composable. So this value is always constant and when we check for
it we always read the default value set at the beginning of app launch.

This PR changes this flow to directly read the toggle value when we get the callback. It also fixes the callback setter
so that we set the toggle first and then call the callback, since the callback is onToggleChanged.