Try clicking some items twice in instrumentation tests (#768)
Despite my best attempts, clicks will sometimes fail to be delivered to the target Composable -- regardless of how many assertions/matchers/polls I do.  To get CI to fail less often, add a new `flakyClickOnNode` function that will try to perform the click twice, with the second click only being triggered if the first click doesn't do what we expected.  This function is applied to the most notoriously flaky clickables, including:
* Clicking on the URL bar to start editing
* Opening the `CardGrid`
* Clicking on "clear" when trying to delete text from the `AutocompleteTextEdit`
* Clicking on the web browser
* Clicking on the overflow menu

I've added a TODO to try and figure out how to make these more stable, but given how these particular clicks are on the critical path for most tests, it at least avoids having to start the whole run of tests again.

This shows up in the logcat in a run that would have otherwise failed like this:
```
06-29 22:28:28.219  5693  5709 I TestRunner: started: createRegularChildTab(com.neeva.app.browsing.LongPressChildTabBehaviorTest)
...
06-29 22:28:47.613  1843  1902 I GnssLocationProvider: WakeLock acquired by sendMessage(REPORT_SV_STATUS, 0, com.android.server.location.GnssLocationProvider$SvStatusInfo@e71c979)
06-29 22:28:47.614  1843  1857 I GnssLocationProvider: WakeLock released by handleMessage(REPORT_SV_STATUS, 0, com.android.server.location.GnssLocationProvider$SvStatusInfo@e71c979)
06-29 22:28:47.890  5693  5709 E ComposeTestUtils: Failed to click using TestTag = 'LocationLabel''.  Trying again.
06-29 22:28:48.613  1843  1902 I GnssLocationProvider: WakeLock acquired by sendMessage(REPORT_SV_STATUS, 0, com.android.server.location.GnssLocationProvider$SvStatusInfo@e6bd6be)
06-29 22:28:48.614  1843  1857 I GnssLocationProvider: WakeLock released by handleMessage(REPORT_SV_STATUS, 0, com.android.server.location.GnssLocationProvider$SvStatusInfo@e6bd6be)
...
06-29 22:28:56.489  5693  5709 I TestRunner: finished: createRegularChildTab(com.neeva.app.browsing.LongPressChildTabBehaviorTest)
06-29 22:28:56.498  5693  5709 I TestRunner: run finished: 1 tests, 0 failed, 0 ignored
```