Persist timestamps for tabs that don't have one persisted on restore (#895)
If a `Tab` doesn't have a timestamp, then we default to assigning it a timestamp of right now.  We currently don't, however, persist this back to the Tab's data via WebLayer, meaning that any tabs that weren't manually selected by the user will get updated with another "now" timestamp when the app is reopened that the tab will never age.

Instead, add a step to the browser restoration that checks if the timestamp was updated to "now" during restoration.  If so, it will get immediately re-persisted.