Use stable sort for tabs displayed in the grid (#899)
Instead of using a chronological reverse sort based on when the tabs in each tab section were last active, use WebLayer's ordering.

Also, add a feature flag that allows the items to be reversed in each bucket.  This will allow items to stay bucketed in their sections with more newly created tabs floating to the top of the list.  This would help with tabs getting completely lost in the "Last 7 days" section, where failing to do a reverse sort would send the newest older tabs all the way to the bottom of that entire section.

Also tries to avoid another way that https://github.com/neevaco/neeva-android/issues/878 could happen, which was causing all the tests to be flaky.