Make the archived tab list use the database instead (#934)
At startup: automatically close any tabs older than the user's archive limit (7 days or 30 days). These are written into the archived tab table. This means that the chronological TabGrid no longer needs to figure out which items are archived and split the list in two (one for the ChronologicalTabGrid and one for the ArchivedTabList).

The Archived Tabs UI now shows tabs from the TabData table. This means that Archived Tabs will only ever show tabs that were closed for the regular profile and never for incognito.

Clicking on an Archived tab "restores" the tab, removing it from the "Archived tab" UI and creating a new tab with the last known URL of the archived tab.