Add settings UI to enable Ad blocker (#955)
This PR adds an Ad blocking toggle below the tracking prevention mode
selection. This toggle is only available when in Strict mode. The code
enforces this by disabling the Ad blocking toggle when Standard mode
is selected.

The copy is updated to better explain the options and make it so that ad
blocking fits.

Issue #954