Force the app to enter Browser mode instead of going to the tab switcher (#1087)
After https://github.com/neevaco/neeva-android/pull/1068/, we can get into a situation where the browser died in the background and sends the user to the tab switcher instead of the tab that is being created for the link in the external VIEW Intent.  Force the user to go to the browser in situations where it's obvious where there's a tab to switch to.