Work around about:blank tabs being created (#1091)
When the user clicks on a link and it opens a new tab, it's possible for WebLayer to end up firing an Intent to open the link in another app.  When this happens, WebLayer automatically closes the tab but it sets the active tab back to null.  This is problematic because WebLayer sends the user to about:blank when they come back to our app.

To work around this, detect when the user is coming back to our app without an active tab set and set it to whatever the most recently used tab was.