Fix programmatically closed tabs not selecting new tab (#1118)
As part of https://github.com/neevaco/neeva-android/pull/1068/files, we changed the code to avoid trying to select a new tab if the user was in the tab switcher.  We went too far, however, resulting in the user being stuck on about:blank if the tab closed itself via javascript.  Add the original check back to get the user out of this state.