Fetch user info when the user opens settings (#1159)
* Fetch user info when the user opens settings

Currently, the user's info is only fetched when we detect that the login token has changed, which only happens at startup and when they first log in.  If the fetch fails in either of these scenarios, the user will be unable to properly view their user info in Settings.  Address this by fetching the user's info when they open settings.

* Space them apart

* lint